use qBound to limit value

+++ kstars/fitsviewer/fitsdata.h
@@ -159,7 +159,7 @@ public:
     void setHistogram(FITSHistogram *inHistogram) { histogram = inHistogram; }
 
     // Filter
-    void applyFilter(FITSScale type, float *image=NULL, double min=-1, double max=-1);
+    void applyFilter(FITSScale type, float *image=NULL, float min=-1, float max=-1);
 
     // Rotation counter. We keep count to rotate WCS keywords on save
     int getRotCounter() const;
